requirementts.txt
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
langchain
langchain_openai
langchain_core
openai
load_dotenv
Flask

typing
langchain_community
python-dotenv
langchain-openai
openai
google-search-results
langchain_core
openpyxl
dateparser
